app/requirements.txt
torch 
torchvision
fastapi
pydantic_settings
pillow
faiss-cpu
pandas
polars
tqdm
uvicorn
pyyaml
python-multipart
numpy
